requirements.txt
jupyter
google-generativeai
streamlit
python-dotenv